chore: merge locators in a single file
Currently, the locators folder is structures in such
way that it leads to cyclic dependencies that do
not work safely in all enviroments. This PR merges
locators in a single file to enable cycle detection
lint rule. Afterwards, we can move it to separate files
once the cycle issue is resolved.
